+++ processor/src/test/resources/expected/com/example/prop/React4j_ObservableProp.java
@@ -1,8 +1,11 @@
-package com.example.arez;
+package com.example.prop;
 
-import arez.Disposable;
+import arez.ObservableValue;
 import arez.annotations.Action;
 import arez.annotations.ArezComponent;
+import arez.annotations.Observable;
+import arez.annotations.ObservableValueRef;
+import java.util.Objects;
 import javax.annotation.Generated;
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
@@ -17,62 +20,49 @@ import react4j.ReactConfig;
 import react4j.arez.ReactArezConfig;
 
 @ArezComponent(
-    name = "ComponentWithDependency"
+    name = "ObservableProp"
 )
 @Generated("react4j.processor.ReactProcessor")
-abstract class React4j_ComponentWithDependency extends ComponentWithDependency {
+abstract class React4j_ObservableProp extends ObservableProp {
   static final ComponentConstructorFunction TYPE = getConstructorFunction();
 
   static final String PROP_value = ReactConfig.shouldMinimizePropKeys() ? "a" : "value";
-
-  static final String PROP_model = ReactConfig.shouldMinimizePropKeys() ? "b" : "model";
 
   @Nonnull
   private static ComponentConstructorFunction getConstructorFunction() {
     final ComponentConstructorFunction componentConstructor = ReactArezConfig.shouldStoreArezDataAsState() ? NativeReactComponent::new : LiteNativeReactComponent::new;
     if ( ReactConfig.enableComponentNames() ) {
-      Js.asPropertyMap( componentConstructor ).set( "displayName", "ComponentWithDependency" );
+      Js.asPropertyMap( componentConstructor ).set( "displayName", "ObservableProp" );
     }
     return componentConstructor;
   }
 
   @Override
-  protected final boolean anyPropsDisposed() {
-    final ComponentWithDependency.Model $$react4jv$$_getModel = getModel();
-    if ( Disposable.isDisposed( $$react4jv$$_getModel ) ) {
-      return true;
-    }
-    return false;
-  }
-
-  @Override
-  protected String getValue() {
+  @Observable(
+      name = "value",
+      expectSetter = false,
+      readOutsideTransaction = true
+  )
+  protected Object getValue() {
     if ( ReactConfig.shouldCheckInvariants() ) {
-      return null != props().getAny( PROP_value ) ? props().getAny( PROP_value ).asString() : null;
+      return null != props().getAny( PROP_value ) ? props().getAny( PROP_value ).cast() : null;
     } else {
       return Js.uncheckedCast( props().getAny( PROP_value ) );
     }
   }
 
-  @Override
-  protected ComponentWithDependency.Model getModel() {
-    if ( ReactConfig.shouldCheckInvariants() ) {
-      return null != props().getAny( PROP_model ) ? props().getAny( PROP_model ).cast() : null;
-    } else {
-      return Js.uncheckedCast( props().getAny( PROP_model ) );
-    }
-  }
+  @Nonnull
+  @ObservableValueRef
+  protected abstract ObservableValue getValueObservableValue();
 
   @Override
   @Action(
       verifyRequired = false
   )
-  protected boolean shouldComponentUpdate(@Nullable final JsPropertyMap<Object> nextProps) {
+  protected boolean notifyOnObservablePropChanges(@Nullable final JsPropertyMap<Object> nextProps) {
     boolean modified = false;
-    if ( !Js.isTripleEqual( props().get( PROP_value ), null == nextProps ? null : nextProps.get( PROP_value ) ) ) {
-      modified = true;
-    }
-    if ( !Js.isTripleEqual( props().get( PROP_model ), null == nextProps ? null : nextProps.get( PROP_model ) ) ) {
+    if ( !Objects.equals( props().get( PROP_value ), null == nextProps ? null : nextProps.get( PROP_value ) ) ) {
+      getValueObservableValue().reportChanged();
       modified = true;
     }
     return modified;
@@ -107,15 +97,15 @@ abstract class React4j_ComponentWithDependency extends ComponentWithDependency {
         @Nonnull JsPropertyMap<Object> arg1);
   }
 
-  private static final class LiteNativeReactComponent extends NativeAdapterComponent<ComponentWithDependency> implements LiteLifecycle {
+  private static final class LiteNativeReactComponent extends NativeAdapterComponent<ObservableProp> implements LiteLifecycle {
     @JsConstructor
     LiteNativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
       super( props );
     }
 
     @Override
-    protected ComponentWithDependency createComponent() {
-      return new Arez_React4j_ComponentWithDependency();
+    protected ObservableProp createComponent() {
+      return new Arez_React4j_ObservableProp();
     }
 
     @Override
@@ -130,15 +120,15 @@ abstract class React4j_ComponentWithDependency extends ComponentWithDependency {
     }
   }
 
-  private static final class NativeReactComponent extends NativeAdapterComponent<ComponentWithDependency> implements Lifecycle {
+  private static final class NativeReactComponent extends NativeAdapterComponent<ObservableProp> implements Lifecycle {
     @JsConstructor
     NativeReactComponent(@Nullable final JsPropertyMap<Object> props) {
       super( props );
     }
 
     @Override
-    protected ComponentWithDependency createComponent() {
-      return new Arez_React4j_ComponentWithDependency();
+    protected ObservableProp createComponent() {
+      return new Arez_React4j_ObservableProp();
     }
 
     @Override