+++ processor/src/test/resources/input/com/example/prop/ObservableViaMemoizeProp.java
@@ -1,0 +1,29 @@
+package com.example.prop;
+
+import arez.annotations.Memoize;
+import javax.annotation.Nullable;
+import react4j.ReactNode;
+import react4j.annotations.Prop;
+import react4j.annotations.ReactComponent;
+import react4j.arez.ReactArezComponent;
+
+@ReactComponent
+abstract class ObservableViaMemoizeProp
+  extends ReactArezComponent
+{
+  @Prop
+  protected abstract Object getValue();
+
+  @Memoize
+  int someValue( int i )
+  {
+    return 0;
+  }
+
+  @Nullable
+  @Override
+  protected ReactNode render()
+  {
+    return null;
+  }
+}